requirements.txt
imgviz==1.2.6
mahotas==1.4.13
numpy
opencv_python
Pillow
PyQt5
PyYAML
torch
torchvision
pycocotools
matplotlib
requests
flask   # only server need this.
# onnxruntime
# onnx